requirements.txt
pandas
numpy
scikit-learn
xgboost
fastapi
uvicorn
pydantic
joblib
python-multipart
django
django-cors-headers
whitenoise
python-dotenv
PyJWT
psycopg2-binary
daphne
channels
google-cloud-aiplatform>=1.38.0
drf-spectacular
djangorestframework
djangorestframework_simplejwt
pytest 
pytest-django
django-unique-user-email
matplotlib
seaborn
plotly
plotly-express
pytrends
scipy
nbformat
requests
shap
django-ratelimit
channels